fix: apply headers for prerendered routes
